+++ MarvelUniverse/app/build.gradle
@@ -34,7 +34,10 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 
+    //Material Design
     implementation 'com.google.android.material:material:1.2.1'
+
+
