+++ MarvelUniverse/app/src/main/AndroidManifest.xml
@@ -2,14 +2,19 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.besugos.marveluniverse">
 
+
+    <uses-permission android:name="android.permission.INTERNET"/>
+
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme">
-        <activity android:name=".SplashActivity" android:theme="@style/Splash">
+        android:theme="@style/AppTheme"
+        >
+        <activity android:name=".splash.view.SplashActivity" android:theme="@style/Splash">
         <intent-filter>
             <action android:name="android.intent.action.MAIN" />
 
@@ -19,7 +24,7 @@
         <activity android:name=".MainActivity">
 
         </activity>
-        <activity android:name=".LoginActivity"/>
+        <activity android:name=".login.view.LoginActivity"/>
     </application>
 
 </manifest>
