+++ MarvelUniverse/app/src/main/java/com/besugos/marveluniverse/login/view/LoginActivity.kt
@@ -1,7 +1,8 @@
-package com.besugos.marveluniverse
+package com.besugos.marveluniverse.login.view
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.besugos.marveluniverse.R
 
 class LoginActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
