+++ MarvelUniverse/app/src/main/java/com/besugos/marveluniverse/splash/view/SplashActivity.kt
@@ -1,0 +1,23 @@
+package com.besugos.marveluniverse.splash.view
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
+import com.besugos.marveluniverse.login.view.LoginActivity
+import com.besugos.marveluniverse.R
+
+class SplashActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        setContentView(R.layout.activity_splash)
+
+        Handler(Looper.getMainLooper()).postDelayed({
+            val intent = Intent(this, LoginActivity::class.java)
+            startActivity(intent)
+            finish()
+        }, 2000)
+    }
+}
