+++ MarvelUniverse/app/src/main/res/layout/activity_splash.xml
@@ -6,22 +6,22 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@color/colorPrimary"
-    tools:context=".SplashActivity">
+    tools:context=".splash.view.SplashActivity">
 
     <ImageView
         android:layout_width="match_parent"
-        android:layout_height="180dp"
-        android:layout_marginTop="230dp"
+        android:layout_height="@dimen/dim_180_dp"
+        android:layout_marginTop="@dimen/dim_230_dp"
         android:src="@drawable/marvel"/>
 
     <TextView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="UNIVERSE"
+        android:text="@string/universe"
         android:textColor="@android:color/white"
         android:textAlignment="center"
         android:textStyle="bold"
-        android:layout_marginTop="-30dp"
+        android:layout_marginTop="@dimen/dim_minus_30_dp"
         android:textSize="@dimen/dim_60_sp"
         />
