+++ MarvelUniverse/app/src/main/res/values/colors.xml
@@ -5,4 +5,7 @@
     <color name="colorAccent">#03DAC5</color>
     <color name="colorMarvelRed">#EE171F</color>
     <color name="colorLightGray">#EFF2F7</color>
+    <color name="facebookBlue">#1875EB</color>
+    <color name="googleBlue">#3883FD</color>
+    <color name="colorTrans">#00EFF2F7</color>
 </resources>
